Separate encoders & decoders into their own files from oracles
