+++ ethereum_history_api/oracles/src/noir/oracles/accountOracle/fixtures/stateProof.json
