+++ ethereum_history_api/oracles/src/noir/oracles/headerOracle.ts
@@ -1,11 +1,10 @@
 import { type ForeignCallOutput } from '@noir-lang/noir_js';
-import { type BlockHeader, blockToHeader, headerToRlp } from '../../ethereum/blockHeader.js';
-import { padArray } from '../../util/array.js';
-import { type GetBlockReturnType, hexToBytes, isHex, keccak256, type PublicClient } from 'viem';
+import { type BlockHeader, blockToHeader } from '../../ethereum/blockHeader.js';
+import { type GetBlockReturnType, isHex, type PublicClient } from 'viem';
 import { assert } from '../../util/assert.js';
 import { NoirArguments } from './oracles.js';
 import { decodeField } from './codec/decode.js';
-import { encodeHex, encodeField } from './codec/encode.js';
+import { encodeBlockHeader } from './headerOracle/codec.js';
 
 export const MAX_HEADER_RLP_SIZE = 708;
 
@@ -26,28 +25,4 @@ export function parseNoirGetHeaderArguments(args: NoirArguments): bigint {
 export async function getBlock(client: PublicClient, blockNumber: bigint): Promise<BlockHeader> {
   const block: GetBlockReturnType = (await client.getBlock({ blockNumber })) as GetBlockReturnType;
   return blockToHeader(block);
-}
-
-export function encodeBlockHeader(header: BlockHeader): ForeignCallOutput[] {
-  return [...encodeBlockHeaderPartial(header), ...encodeBlockHeaderRlp(header)];
-}
-
-function encodeBlockHeaderPartial(header: BlockHeader): ForeignCallOutput[] {
-  const rlpHex = headerToRlp(header);
-
-  const number = header.number;
-  const hash = encodeHex(keccak256(hexToBytes(rlpHex)));
-  const stateRoot = encodeHex(header.stateRoot);
-  const transactionsRoot = encodeHex(header.transactionsRoot);
-  const receiptsRoot = encodeHex(header.receiptsRoot);
-  return [number, hash, stateRoot, transactionsRoot, receiptsRoot];
-}
-
-function encodeBlockHeaderRlp(header: BlockHeader): ForeignCallOutput[] {
-  const rlpHex = headerToRlp(header);
-  const rlpBytes = encodeHex(rlpHex);
-
-  const encodedRlpLen = encodeField(rlpBytes.length);
-  const encodedRlp = padArray(rlpBytes, MAX_HEADER_RLP_SIZE, '0x');
-  return [encodedRlpLen, encodedRlp];
 }
