+++ ethereum_history_api/oracles/tsconfig.json
@@ -1,6 +1,6 @@
 {
   "extends": "../../tsconfig.common.json",
-  "include": ["src/**/*", "src/noir/oracles/fixtures/*.json", "package.json"],
+  "include": ["src/**/*", "src/noir/oracles/accountOracle/fixtures/*.json", "package.json"],
   "compilerOptions": {
     "rootDirs": ["src"],
     "outDir": "dist"
